+++ tatk/policy/loader/dataset.py
@@ -1,0 +1,15 @@
+import torch.utils.data as data
+
+class ActDataset(data.Dataset):
+    def __init__(self, s_s, a_s):
+        self.s_s = s_s
+        self.a_s = a_s
+        self.num_total = len(s_s)
+    
+    def __getitem__(self, index):
+        s = self.s_s[index]
+        a = self.a_s[index]
+        return s, a
+    
+    def __len__(self):
+        return self.num_total
